% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_adfgSubs-package.R
\docType{package}
\name{adfgSubs-package}
\alias{adfgSubs}
\alias{adfgSubs-package}
\title{ADF&G Subsistence Data Analysis Functions}
\description{
Functions for analysis of ADF&G Subsistence survey data.
}
\section{Installation}{

To install the package without updating dependencies (recommended):

\code{devtools::install_github("ADFG-Subsistence-Division/adfgSubs", upgrade = "never")}
}

\section{Package Information}{

\itemize{
  \item Last Updated: 2025-03-19 19:53:21 UTC
  \item Maintainer: jmcoleman2-sub-ak
  \item Version: 0.1.0
}
}

\author{
\strong{Maintainer}: Coleman Jesse \email{jesse.coleman@alaska.gov}

Authors:
\itemize{
  \item Koster David \email{david.koster@alaska.gov}
}

}
\keyword{internal}
