% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{assign_value_variables}
\alias{assign_value_variables}
\title{assign_value_variables}
\usage{
assign_value_variables(dataSetA, variableList, newValue)
}
\arguments{
\item{dataSetA}{A data frame containing multiple variables.}

\item{variableList}{A vector containing the variables to be recoded.}

\item{newValue}{The value that will be assigned to variables specified in @param variableList.}
}
\value{
A modified data frame with new values in variables specified by @param variableList.
}
\description{
Replace whatever is in the specified columns with the assigned value, without regard for the current contents.
}
