% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_summarizeSubsData.R
\name{calculateCI}
\alias{calculateCI}
\title{calculateCI}
\usage{
calculateCI(x, conf_level = 0.95, N = NULL)
}
\arguments{
\item{x}{An estimated variable for which confidence intervals are to be created.}

\item{conf_level}{The desired confidence level for the estimate.}

\item{N}{The total number households in the community.}
}
\value{
The lower CI, the upper CI, and the
}
\description{
calculateCI
}
