% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_summarizeSubsData.R
\name{calculateMode}
\alias{calculateMode}
\title{calculateMode}
\usage{
calculateMode(x)
}
\arguments{
\item{x}{A numeric variable for which the mode will be calculated.}
}
\value{
The mode of the values in the supplied variable.
}
\description{
Helper function
}
