% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_combineCI.R
\name{combineCI}
\alias{combineCI}
\title{Combine multiple estimates and confidence intervals}
\usage{
combineCI(est, lower, upper, conf_level = 0.95)
}
\arguments{
\item{est}{A column containing the estimates to be combined}

\item{lower}{A column containing the lower confidence bounds to be combined}

\item{upper}{A column containing the upper confidence bounds to be combined}

\item{conf_level}{The desired confidence level for the combined interval (default is 0.95)}
}
\value{
A tibble with combined estimate and confidence interval
}
\description{
Creates a combined estimate and confidence interval from individual estimates and their confidence intervals.
}
\examples{
# Example usage:
}
