% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToBBPermit}
\alias{connectToBBPermit}
\title{connectToBBPermit}
\usage{
connectToBBPermit(server)
}
\arguments{
\item{server}{Server name.}
}
\value{
Server connection.
}
\description{
Initializes connection to Bristol Bay subsistence salmon fishery permit database server.
}
