% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToCSHDB}
\alias{connectToCSHDB}
\title{connectToCSHDB}
\usage{
connectToCSHDB(server)
}
\arguments{
\item{server}{Server name.}
}
\value{
Server connection.
}
\description{
Initializes connection to community subsistence hunt (GMU 13 moose and caribou) database server.
}
