% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToCSIS}
\alias{connectToCSIS}
\title{connectToCSIS}
\usage{
connectToCSIS()
}
\value{
Server connection.
}
\description{
Initializes connection to Community Subsistence Information System server.
}
