% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToHalibut}
\alias{connectToHalibut}
\title{connectToHalibut}
\usage{
connectToHalibut()
}
\value{
Server connection.
}
\description{
Initializes connection to halibut database server.
}
