% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToHHDB}
\alias{connectToHHDB}
\title{connectToHHDB}
\usage{
connectToHHDB()
}
\value{
Server connection.
}
\description{
Initializes connection to household database server.
}
