% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToKuskokwim}
\alias{connectToKuskokwim}
\title{connectToKuskokwim}
\usage{
connectToKuskokwim()
}
\arguments{
\item{server}{Server name.}
}
\value{
Server connection.
}
\description{
Initializes connection to Kusko database server.
}
