% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{delete_variables}
\alias{delete_variables}
\title{delete_variables}
\usage{
delete_variables(dataSetD, variablesToDelete)
}
\arguments{
\item{dataSetD}{A data frame containing multiple variables.}

\item{variablesToDelete}{A vector containing the variables to be deleted from @param dataSetD.}
}
\value{
A modified data frame with variables named in @param variablesToDelete removed.
}
\description{
Removes unnecessary variables from a data frame.

Replicates SPSS functions for:
  'DELETE VARIABLES'
   /KEEP (when saving)
  'RENAME VARIABLES'
  'RECODE x .. y (a = b).'
}
