% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{errorMessage}
\alias{errorMessage}
\title{errorMessage}
\usage{
errorMessage(fmtText = "")
}
\arguments{
\item{fmtText}{}
}
\value{
Formatted error messages.
}
\description{
Format error messages
}
