% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{figCaption}
\alias{figCaption}
\title{figCaption}
\usage{
figCaption(caption = "Plot")
}
\arguments{
\item{caption}{}
}
\value{
Formatted figure captions.
}
\description{
Format figure captions
}
