% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{formatSummaryBlock}
\alias{formatSummaryBlock}
\title{formatSummaryBlock}
\usage{
formatSummaryBlock(fmtTxt = "")
}
\arguments{
\item{fmtTxt}{}
}
\value{
Formatted summary blocks.
}
\description{
Format summary blocks
}
