% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{formatTableHeader}
\alias{formatTableHeader}
\title{formatTableHeader}
\usage{
formatTableHeader(fmtTxt = "")
}
\arguments{
\item{fmtTxt}{}
}
\value{
Formatted table headers.
}
\description{
Format table headers
}
\details{
These rely on the style.css developed for each project. To make changes to the style, please modify style.css
}
