% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{formatValueList}
\alias{formatValueList}
\title{formatValueList}
\usage{
formatValueList(fmtTxt = c())
}
\arguments{
\item{fmtTxt}{}
}
\value{
Formatted values lists.
}
\description{
Format values lists
}
