% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{get_code_labels}
\alias{get_code_labels}
\title{get_code_labels}
\usage{
get_code_labels(codesToFind, codeColumnName, valueColumnName, codeList)
}
\arguments{
\item{codesToFind}{Description}

\item{codeColumnName}{Description}

\item{valueColumnName}{Description}

\item{codeList}{Description}
}
\value{
Description
}
\description{
Description
}
