% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plottingFunctions.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(palName = "Spectral", nColors = 1, override = FALSE)
}
\arguments{
\item{palName}{Palette name; default is "Spectral"}

\item{nColors}{Number of colors needed}

\item{override}{Logical; should colorblind-safe values be used? Default is FALSE}
}
\value{
A vector of color values.
}
\description{
Helper function for color palettes. This function provides a way for
 analysts to do a better job of customizing colors per-project,
 while starting with the default scheme.
}
