% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_getRecName.R
\name{getRecName}
\alias{getRecName}
\title{Extract the record name from the current Rmd}
\usage{
getRecName()
}
\value{
The name of the record (e.g., "REC04")
}
\description{
Extract the record name from the current Rmd
}
