% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_summarizeSubsData.R
\name{getSubsStratifiedSurveyEstimate}
\alias{getSubsStratifiedSurveyEstimate}
\title{getSubsStratifiedSurveyEstimate}
\usage{
getSubsStratifiedSurveyEstimate(
  data,
  summarize_vars,
  grouping_vars = c("projID", "studyear", "communty", "strata"),
  N_col = "commhh",
  strata_col = "strata"
)
}
\arguments{
\item{x}{A list of variables for which expanded estimates, confidence intervals,
and relative margins of error will be calculated and summarized.}
}
\value{
Estimates, confidence intervals, and relative margins of error
}
\description{
Calculates variable estimates for stratified samples.
}
