% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_summarizeSubsData.R
\name{getSubsSurveyEstimate}
\alias{getSubsSurveyEstimate}
\title{getSubsSurveyEstimate}
\usage{
getSubsSurveyEstimate(
  data,
  summarize_vars,
  grouping_vars = c("projID", "studyear", "communty", "strata"),
  N_col = "commhh",
  addvariables = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{summarize_vars}{A vector containing names of variables to be summarized.}

\item{grouping_vars}{A vector containing names of variables to group by;
one estimate is given for each combination of grouping variables.}

\item{N_col}{A column in the data frame with the total number of households (N).}

\item{addvariables}{Logical; should the columns containing descriptive statistics
and the estimate be returned in addition to the summarized data frame? Default is 'FALSE'.}
}
\value{
A data frame with summarized variables.
}
\description{
Summary data specific to subsistence the analysis design for subsistence.
}
