% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_GGPLOT_FUNCTIONS.R
\name{ggBoxWhiskerFormatted}
\alias{ggBoxWhiskerFormatted}
\title{ggBoxWhiskerFormatted}
\usage{
ggBoxWhiskerFormatted(
  figData,
  categoryColName,
  dataColName,
  yAxisTitle = "Reported number",
  yBreaks = 20
)
}
\arguments{
\item{figData}{description}

\item{categoryColName}{description}

\item{dataColName}{description}

\item{yAxisTitle}{default = "Reported number"}

\item{yBreaks}{default = 20}
}
\value{
description
}
\description{
Function for producing ggplot outputs according to current ADF&G Division of Subsistence Subsistence standards.
}
\details{
Additional details...
These functions rely on ggplot2 and standardized inputs. This file
}
