% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_giniCoefficient.R
\name{gini_lorenz}
\alias{gini_lorenz}
\title{gini_lorenz}
\usage{
gini_lorenz(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The Gini coefficient.
}
\description{
Calculates the Gini coefficient of a series of values using the Lorenz curve area method.
}
\details{
Citation: Sen, A. 1997. On Economic Inequality. Clarendon Press.
}
