% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_giniCoefficient.R
\name{gini_mad}
\alias{gini_mad}
\title{gini_mad}
\usage{
gini_mad(y)
}
\arguments{
\item{y}{A vector of ranked harvest values.}
}
\value{
The Gini coefficient.
}
\description{
Calculates the Gini coefficient of a series of values using the mean absolute difference method.
}
\details{
Citation: Cowell, F. A. 2011. Measuring Inequality. 3rd edition. Oxford University Press. Gini coefficient formula on page 155.
}
