% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{greenMessage}
\alias{greenMessage}
\title{greenMessage}
\usage{
greenMessage(fmtText = "")
}
\arguments{
\item{fmtText}{}
}
\value{
Formatted green messages.
}
\description{
Format green messages
}
