% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{inf_to_NA}
\alias{inf_to_NA}
\title{inf_to_NA}
\usage{
inf_to_NA(dataSetI, variableList)
}
\arguments{
\item{dataSetI}{A data frame containing multiple variables.}

\item{variableList}{A vector containing the variables to be recoded.}
}
\value{
A modified data frame with 'inf' values recoded to 'NA' in variables specified by `variableList`.
}
\description{
Recodes variables with 'inf' values to 'NA'.
}
