% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_insertChunkAfter.R
\name{insertChunkAfter}
\alias{insertChunkAfter}
\title{Insert a new chunk after a specified code chunk in Rmd files}
\usage{
insertChunkAfter(
  path = ".",
  pattern = "\\\\.Rmd$",
  after_chunk_label,
  new_header = NULL,
  new_text = NULL,
  new_chunk = NULL
)
}
\arguments{
\item{path}{Character. Path to directory containing files to process. Default is current directory (\code{"."}).}

\item{pattern}{Character. Regex pattern to match files. Default is \code{"\\.Rmd$"}.}

\item{after_chunk_label}{Character. The label of the code chunk after which the new content should be inserted (e.g., \code{"setup"} for a chunk labeled \code{r setup}).}

\item{new_header}{Character vector or NULL. Optional. Section header(s) to insert after the chunk (e.g., \code{"## New Step"}).}

\item{new_text}{Character vector or NULL. Optional. Descriptive text to insert after the header.}

\item{new_chunk}{Character vector or NULL. Optional. Lines of a new code chunk to insert.}
}
\value{
Invisibly returns NULL. Used for its side effect of editing files in place.
}
\description{
This function searches R Markdown (\code{.Rmd}) files in a specified directory and inserts new content (header, text, and/or an R chunk)
immediately after a code chunk with a specified label.
}
\examples{

#Example usage:

newHeader <- c("## Conversion factors for unspecified resources")
newText <- c(
  "Resources are often reported as 'unspecified', in cases where the survey respondent",
  "was unsure of the species harvested or received. To calculate the edible weight of",
  "unspecified resources harvested, similar resources and the amounts/weights harvested",
  "are used to calculate the conversion factor for unspecified resources. Analysts must",
  "develop a list of resources to use in this calculation for each unspecified resource."
)
newChunk <- c()
insertChunkAfter(
  path = "analysis",
  after_chunk_label = "standardize units",
  new_header = newHeader,
  new_text = newText,
  new_chunk = newChunk
)

}
