% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_kblSubsTabbed.R
\name{kblSubsTabbed}
\alias{kblSubsTabbed}
\title{Outputs R Markdown tabset syntax with a table per group}
\usage{
kblSubsTabbed(df, groupCol, tableFun = kableExtra::kbl, ...)
}
\arguments{
\item{df}{Data frame}

\item{groupCol}{Name or position of grouping variable}

\item{tableFun}{Function to render table (default: kableExtra::kbl)}

\item{...}{Passed to tableFun}
}
\description{
Outputs R Markdown tabset syntax with a table per group
}
\examples{
kblSubsTabbed(iris, "Species", kable::kable)
}
