% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{keep_variables}
\alias{keep_variables}
\title{keep_variables}
\usage{
keep_variables(dataSetK, variablesToKeep)
}
\arguments{
\item{dataSetK}{A data frame containing multiple variables.}

\item{variablesToKeep}{A vector containing the variables to be retained from `dataSetK`. All other variables will be deleted.}
}
\value{
A modified data frame with only variables named in `variablesToKeep` retained.
}
\description{
Retains only specified variables from a data frame.
}
