% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_makeLabelShapes.R
\name{makeLabelShapes}
\alias{makeLabelShapes}
\title{Assign Plot Shapes to Labels Based on Pattern Matching}
\usage{
makeLabelShapes(labels, patterns, shapes)
}
\arguments{
\item{labels}{A character vector of labels to be assigned shapes.}

\item{patterns}{A character vector of regular expressions (Perl-compatible) to
match against labels. Matching is performed sequentially, and the first match
determines the assigned shape.}

\item{shapes}{A character or integer vector of plot shape values corresponding
to each pattern. Values will be coerced to integers. Must have the same length
as \code{patterns}.}
}
\value{
A named integer vector where names are the input \code{labels} and
  values are the assigned plot shapes. Labels with no matching pattern are
  assigned \code{NA_integer_}.
}
\description{
This function maps categorical labels to plot shapes (as integers) by matching
each label against a series of regex patterns and returning the shape associated
with the first matching pattern. Useful for creating consistent shape schemes in
visualizations where shapes should be assigned based on label characteristics.
}
\details{
Pattern matching is performed sequentially using \code{\link[base]{grepl}} with
Perl-compatible regular expressions. The function returns the shape associated
with the first pattern that matches a given label. If a label matches multiple
patterns, only the first match is used.

Shape values are standard ggplot2 point shapes (0-25), where common values include:
16 (filled circle), 17 (filled triangle), 18 (filled diamond), 15 (filled square).
}
\examples{
labels <- c("harvest_HH2024", "survey_data", "harvest_HH2025", "other_file")
patterns <- c("^harvest", "^survey")
shapes <- c(16, 17)

makeLabelShapes(labels, patterns, shapes)

}
