% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{nan_to_NA}
\alias{nan_to_NA}
\title{nan_to_NA}
\usage{
nan_to_NA(datasetN, variableList)
}
\arguments{
\item{variableList}{A vector containing the variables to be recoded.}

\item{dataSetN}{A data frame containing multiple variables.}
}
\value{
A modified data frame with 'NaN' values recoded to 'NA' in variables specified by `variableList`.
}
\description{
Recodes variables with 'NaN' values to 'NA'.
}
