% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plottingFunctions.R
\name{pieSubs}
\alias{pieSubs}
\title{adfgSubs themed pie charts}
\usage{
pieSubs(
  df,
  value_col = "value",
  percent_col = "percent",
  commname_col = "commname",
  category_col = "category",
  label_col = "label",
  colors = NULL,
  main = NULL,
  font_family = "Times New Roman",
  font_size = 1,
  leader_line_col = "gray50",
  leader_line_len = 1.2
)
}
\arguments{
\item{df}{A data frame containing variables to be plotted.}

\item{value_col}{A column of values to be represented by pie slices.}

\item{percent_col}{A column of percentage values to be represented}

\item{commname_col}{A column of community name values.}

\item{category_col}{A column of resource category values to be mapped to colors.}

\item{label_col}{A column of labels combining resource category/percentage.}

\item{colors}{A vector of colors to be used for plotting.}

\item{main}{The main title of the plot.}

\item{font_family}{Font family; default is "serif".}

\item{font_size}{Font size; default is 12 point.}

\item{leader_line_col}{Color of the leader lines between labels and pie slices.}

\item{leader_line_len}{Length of the leader lines between labels and pie slices.}
}
\value{
A pie chart.
}
\description{
A function to create custom pie charts with base R. This is an alternative to
ggplot's coord_polar() transformation, which doesn't always lend itself to
correct label placement.
}
