% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manageVariables.r
\name{recode_variables}
\alias{recode_variables}
\title{recode_variables}
\usage{
recode_variables(dataSetR, variableList, oldValue, newValue)
}
\arguments{
\item{dataSetR}{A data frame containing multiple variables.}

\item{variableList}{A vector containing the variables to be recoded.}

\item{oldValue}{Existing value to be replaced.}

\item{newValue}{Value used to replace @param oldValue}
}
\value{
A modified data frame with old values recoded to new values in variables specified by @param variableList.
}
\description{
Recodes variables.
}
