% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_replaceCodeChunk.R
\name{replaceCodeChunk}
\alias{replaceCodeChunk}
\title{Replace the code in a specified chunk in Rmd files}
\usage{
replaceCodeChunk(
  path = ".",
  pattern = "\\\\.Rmd$",
  exclude = NULL,
  chunk_label = NULL,
  new_code = NULL,
  rec_pattern = "REC[0-9]{2}",
  rec_placeholder = "REC_PLACEHOLDER"
)
}
\arguments{
\item{path}{a vector of folder paths containing files where code chunk will be replaced}

\item{pattern}{the file type to be modified; default is ".Rmd"}

\item{exclude}{vector of file names to exclude}

\item{chunk_label}{code chunk label, e.g., "write csv files"}

\item{new_code}{new code to replace existing code in chunk, as a vector of lines (no chunk header/footer)}

\item{rec_pattern}{extracts the record type from the existing code chunk}

\item{rec_placeholder}{placeholder record type vector that is replaced by the extracted record type}
}
\value{
Invisibly returns NULL. Used for its side effect of editing files in place.
}
\description{
Replace the code in a specified chunk in Rmd files
}
\examples{

# Example usage

Can specify full path names or paths relative to root directory.

foldersAmbler <- c(
"../Step 1 - Extraction and Validation",
"../Step 2 - Standardize and Validate",
"../Step 3 - Demographics",
"../Step 4 - Income and Employment",
"../Step 5 - Harvest processing",
"../Step 6 - Assessments and summary")

New code stored as quoted text in vector. Note: any text or characters
 being pasted into chunk should use single quotes.

newCode <- c("# Set some knit options and functions for formatting data.
knitr::opts_chunk$set(echo = FALSE,
                      include = TRUE,
                      message = FALSE,
                      warning = FALSE,
                      results='asis')

options(knitr.kable.NA = '')"
)

replaceCodeChunks(path = foldersAmbler,
chunk_label = "setup",
new_code = newCode)


}
