% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_replaceYAML.R
\name{replaceYAML}
\alias{replaceYAML}
\title{Replace the YAML header in Rmd files}
\usage{
replaceYAML(
  files = NULL,
  folders = NULL,
  pattern = "\\\\.Rmd$",
  author = NULL,
  date = NULL,
  html_lines = NULL,
  output_lines = c("  html_document:", "    theme: null",
    "    css: ../HTML/adfg-style.css", "    include:",
    "      before_body: ../HTML/header.html", "      after_body: ../HTML/footer.html"),
  suffix = NULL
)
}
\arguments{
\item{files}{a vector of .Rmd file paths for which the YAML header will be replaced}

\item{folders}{a vector of .Rmd folder paths for which the YAML header will be replaced}

\item{pattern}{the file type to be modified; default is ".Rmd"}

\item{author}{the name of the author to be updated in the YAML header}

\item{date}{the date to be updated in the YAML header}

\item{html_lines}{extra HTML code to be added below the header for rendering (suffix)}

\item{output_lines}{the text specifying the outupt of the YAML header; defaults to HTML}

\item{suffix}{the project ID and title to be included in the suffix}
}
\value{
Invisibly returns NULL. Used for its side effect of editing files in place.
}
\description{
Replace the YAML header in Rmd files
}
\examples{

# Example usage:

replaceYAML(files = filesKivalinaStep2[c(5:15)]
         author = c("Jesse Coleman"),
         date = c("2025-08-18"),
         suffix = c("- (313) SWG Kivalina Comprehensive"))
}
