% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_summarizeSubsData.R
\name{safeMax}
\alias{safeMax}
\title{Safely return maximum value}
\usage{
safeMax(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\value{
The maximum value from a vector; returns NA if all values are NA.
}
\description{
Safely return maximum value
}
\examples{

library(dplyr)

df <- data.frame(g = c(1,1,2,2), val = c(NA, NA, 2, NA))
df \%>\%
  group_by(g) \%>\%
  summarise(
    my_min = safeMin(val),
    my_max = safeMax(val),
    my_sum = safeSum(val)
  )

# A tibble: 2 × 4
     g      my_min my_max my_sum
     <dbl>  <dbl>  <dbl>  <dbl>
1     1     NA     NA     NA
2     2      2      2      2

}
