% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_standardizeColumns.r
\name{standardizeColumns}
\alias{standardizeColumns}
\title{standardizeColumns}
\usage{
standardizeColumns(sourceData, colList, replaceValue = 0)
}
\arguments{
\item{sourceData}{A data frame.}

\item{colList}{A vector of column names containing variables to be standardized with replacement values.}

\item{replaceValue}{A vector containing a replacement value; default = 0.}
}
\value{
A data frame with new columns containing the replacement value in each row.
}
\description{
Looks through each of the columns in a data frame and if the column isn't present, gives it the value of replaceValue.
}
