% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_standardizeConversionFactors.r
\name{standardizeConvFact}
\alias{standardizeConvFact}
\title{standardizeConvFact}
\usage{
standardizeConvFact(sourceData, convData, variableList)
}
\arguments{
\item{sourceData}{A data frame. Must contain the following columns: "projID", "studyear", "communty", "resource", "units"}

\item{convData}{A data frame containing conversion factors. Must contain the following columns: "projID", "studyear", "communty", "resource", "units", "convFact","defaultUnits"}

\item{variableList}{A vector of variable names for which values will be converted.}
}
\value{
A data frame containing variables converted using specified conversion factors.
}
\description{
Standardized conversion factors function; generalized to be applicable to all harvest record types, provided a given format.
}
