% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_standardizeHHRecords.r
\name{standardizeHHRecords}
\alias{standardizeHHRecords}
\title{standardizeHHRecords}
\usage{
standardizeHHRecords(sourceData, replicationKeyList = c("resource", "units"))
}
\arguments{
\item{sourceData}{A data frame.}

\item{replicationKeyList}{The list of columns that make a row unique for a given
household and project in the supplied data frame. These 
columns may vary depending on project. The default is:
resource, units. This can be overridden, depending on the 
needs of the project.}
}
\value{
A data frame in which all households (rows) have values for every resource column and unit column harvested.
}
\description{
1. Creates a 'template' record for each unique replication key value such that
all replication values are present for each household for a projID/studyear/communty
AND replication key values not appearing in a projID/studyear/communty
     are not added

2. Creates a set of all rows appearing in the dataset and containing the
   replication key, projectKey, and hhKey.

3. Loops Through each household

4. Limits the list of replication keys to just those with matching projID, studyear, communty

5. Adds the HHID and strata value to the template list

6. Uses bind_rows into a placeholder data-frame.
}
\details{
Note:
 Prior to standardizing resource codes, we must
 first have cleaned up files such that there are NO duplicate resource codes.
 NOTE that this code is generic. 

Definitions:
 `replicationKeyList` is the list of columns that make a row unique for a given
 household and project in the supplied data frame. These 
 columns may vary depending on project. The default is:
 resource, units. This can be overridden, depending on the 
 needs of the project.

 `hhKeyList` is the list of columns that specify a unique household for a given
 project in the supplied data frame. These columns are fixed across all
 projects coming from the SDS these are: strata, HHID

 `projKeyList` contains columns that specify a unique project. These columns are 
 fixed across all projects coming from the SDS these are: 
 projID, studyear, communty
}
