% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_themeSubs.R
\name{themeSubs}
\alias{themeSubs}
\title{Apply adfgSubs plot theme}
\usage{
themeSubs(base_size = 10, base_family = "Times New Roman")
}
\arguments{
\item{base_size}{Font size; defaults to 10 point.}

\item{base_family}{Font family; defaults to "Times New Roman".}
}
\value{
No return value. Called for its side effect (applies ggplot formatting)
}
\description{
A ggplot2 theme to mimic tables and figures produced in Excel.
}
