% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_transposeData.r
\name{transposeForTable}
\alias{transposeForTable}
\title{transposeForTable}
\usage{
transposeForTable(ttData, headerCol = NA, roundPrecision = 2)
}
\arguments{
\item{ttData}{A small data frame for producing a table.}

\item{headerCol}{A row id column. Default = NA.}

\item{roundPrecision}{The number of digits to which variables should be rounded. Default = 2.}
}
\value{
A transposed data frame.
}
\description{
Given a data frame, the name of a header column, and rounding precision
(which defaults to 2), this code transposes rows of data into columns.
The data are converted into character values to satisfy R data-requirements.
This code should only be used for organization of display data for
R-markdown or sometimes for output of data files that are presented
in a transposed format, such as the sample tables.
}
