% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{tTestOutput}
\alias{tTestOutput}
\title{tTestOutput}
\usage{
tTestOutput(t.test, title = "t-test results", rounding = TRUE)
}
\arguments{
\item{t.test}{description}

\item{title}{description}

\item{rounding}{description}
}
\value{
Formatted t-test output.
}
\description{
T-test output formatting
}
