% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_verifyColumns.r
\name{verifyColumns}
\alias{verifyColumns}
\title{verifyColumns}
\usage{
verifyColumns(sourceData, colList)
}
\arguments{
\item{sourceData}{A data frame.}

\item{colList}{A vector containing column names to check for.}
}
\value{
Returns an error if columns are missing.
}
\description{
Look through each of the columns and if the column isn't present, return an error message.
}
