% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RMD_HTML_FORMATTING_FUNCTIONS.r
\name{warningMessage}
\alias{warningMessage}
\title{warningMessage}
\usage{
warningMessage(fmtText = "")
}
\arguments{
\item{fmtText}{}
}
\value{
Formatted warning messages.
}
\description{
Format warning messages
}
