% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToASFDB}
\alias{connectToASFDB}
\title{connectToASFDB}
\usage{
connectToASFDB()
}
\arguments{
\item{server}{Server name.}
}
\value{
Server connection.
}
\description{
Initializes connection to Alaska Salmon Fishery Database server.
}
