% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_subsistenceDatabaseConnections.r
\name{connectToSDS}
\alias{connectToSDS}
\title{connectToSDS}
\usage{
connectToSDS()
}
\value{
Server connection.
}
\description{
Initializes connection to Subsistence Database server.
}
