% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_meanReplacement.R
\name{dfMeanReplaceStratified}
\alias{dfMeanReplaceStratified}
\title{Replace missing column values in a data frame}
\usage{
dfMeanReplaceStratified(
  sourceData,
  replCols,
  checkCol = "filterq",
  verbose = TRUE
)
}
\arguments{
\item{sourceData}{A data frame containing multiple variables.}

\item{checkCol}{Description}

\item{verbose}{Logical; should the mean (before and after replacement), number of NA values, and number of zeroes be returned? Default is TRUE}

\item{replCol}{A vector containing the variables from sourceData for which values will be mean replaced.}
}
\value{
A modified data frame containing mean replaced values in the specified variables.
}
\description{
Simple mean replacement for a stratified sample. This function REQUIRES that
   sourceData contain: "projID", "studyear", "communty", "resource",
                       "strata", replCol, "commhh", "NHouseholds".
   additionally, a 'check-column' is required, this might be
   "harvestq" (default) or "filterq" or other specified column to indicate
   whether or not a resource requires a non-zero amount (ie: minimum replacement).
This is to be used for simple circumstances where there are no details
   such as sex, season or gear type.
Note that the inclusion of harvestq in this list is intended to differentiate
  between 'some amount, amount unknown", and instances where the question of
  whether or not a harvest even occurred is unknown. If the status is
  "some amount, amount unknown" i.e., harvestq == 1, then minimum replacement
  occurs, if not then the community mean is supplied regardless.

This function first tries strata-level means:

 Groups by projID, studyear, communty, resource, and strata
 Calculates mean within these fine-grained groups

Then tries community-level means:

 Groups by projID, studyear, communty, and resource
 Uses these broader groups as a secondary fallback
 Provides additional fallbacks:

Finally, if all else fails:
 The minimum value strata is used at the replacement value.

Detailed diagnostics:

Shows values of calculated means at each level
Counts of NA values and zeros
Summary statistics before and after replacement
}
