% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_meanReplacement.R
\name{meanReplaceColumn}
\alias{meanReplaceColumn}
\title{Mean replace values for use with mutate(across(...))}
\usage{
meanReplaceColumn(repl_col, strata, commhh, NHouseholds, checkCol)
}
\arguments{
\item{repl_col}{Numeric vector. The column to be mean replaced (e.g., observed values with possible NAs).}

\item{strata}{Vector (factor, character, or integer). Identifies strata for stratified mean imputation.}

\item{commhh}{Vector. Community household identifier, or grouping used for estimation at the community level.}

\item{NHouseholds}{Numeric or integer vector. The number of households in each group (typically reused per row).}

\item{checkCol}{Numeric or integer vector. Used to determine minimum replacement (1 if >0 and not NA, else 0).}
}
\value{
A numeric vector of the same length as \code{repl_col}, containing the mean-replaced or imputed values.
}
\description{
This function imputes missing values in a numeric vector (e.g., survey data) using stratified means and then community means. If both mean imputations fail, a minimum replacement is used based on a check column. This is typically used to replace missing or invalid values in household-level resource survey data.
}
\examples{

df <- tibble(strata = c("A","A","A","B","B","B"),
             commhh = c(99,99,99,53,53,53),
             NHouseholds = c(99,99,99,53,53,53),
             col1 = c(1, NA, 7, 2, 4, NA),
             col2 = c(18, 39, NA, 109, NA, 44),
             checkCol = c(1,1,1,1,1,1))

replCols <- c("col1","col2")

suffix <- c("MR")

# Usage:
 df \%>\%
    mutate(across(replCols, ~ meanReplaceColumn(.x, strata, commhh, NHouseholds, checkCol = "checkCol"),
    .names = paste0("{.col}_", suffix)))
}
