% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_meanReplacement.R
\name{meanReplaceStratified}
\alias{meanReplaceStratified}
\title{meanReplaceStratified}
\usage{
meanReplaceStratified(sourceData, replCol, checkCol = "harvestq")
}
\arguments{
\item{sourceData}{A data frame containing multiple variables.}

\item{replCol}{A vector containing the variables from sourceData for which values will be mean replaced.}

\item{checkCol}{Description}
}
\value{
A modified data frame containing mean replaced values in the specified variables.
}
\description{
Simple mean replacement for a stratified sample. This function REQUIRES that
   sourceData contain: "projID", "studyear", "communty", "resource",
                       "strata", replCol, "commhh", "NHouseholds".
   additionally, a 'check-column' is required, this might be
   "harvestq" (default) or "filterq" or other specified column to indicate
   whether or not a resource requires a non-zero amount (ie: minimum replacement).
This is to be used for simple circumstances where there are no details
   such as sex, season or gear type.
Note that the inclusion of harvestq in this list is intended to differentiate
  between 'some amount, amount unknown", and instances where the question of
  whether or not a harvest even occurred is unknown. If the status is
  "some amount, amount unknown" ie: harvestq == 1, then minimum replacement
  occurs, if not then the community mean is supplied regardless.
}
