% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_meanReplaceStratified.R,
%   R/f_meanReplacement.R, R/f_replaceByResourceDetail.R
\name{replaceByResourceDetail}
\alias{replaceByResourceDetail}
\title{replaceByResourceDetail}
\usage{
replaceByResourceDetail(sourceData, replCol, detailColumn)

replaceByResourceDetail(sourceData, replCol, detailColumn)

replaceByResourceDetail(sourceData, replCol, detailColumn)
}
\arguments{
\item{sourceData}{A data frame}

\item{replCol}{Column with missing values}

\item{detailColumn}{Column that contains a grouping variable}
}
\value{
A column with missing replaced values (mean or minimum)

A column with missing replaced values (mean or minimum)

A column with missing replaced values (mean or minimum)
}
\description{
This function is for SDS projects only, and assumes the presence of a stratification variable.
}
