% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plottingFunctions.R
\name{themeSubs}
\alias{themeSubs}
\title{themeSubs}
\usage{
themeSubs(base_size = 12, base_family = "serif")
}
\arguments{
\item{base_size}{Font size; defaults to 12 point.}

\item{base_family}{Font family; defaults to "serif".}
}
\value{
Nothing
}
\description{
A ggplot2 theme to mimick tables and figures produced in Excel.
}
